+++ cogstat/cogstat_dialogs.py
@@ -23,23 +23,30 @@ QtCore.QCoreApplication.translate = _gui
 
 def open_data_file():
     return str(QtWidgets.QFileDialog.getOpenFileName(None, _('Open data file'), '',
-                                                     '%s (*.csv *.sav *.zsav *.por *.sas7bdat *.xpt *.dta *.rdata *.rds);;'
+                                                     '%s (*.csv *.sav *.zsav *.por *.sas7bdat *.xpt *.dta *.rdata '
+                                                     '*.rds *.ods *.xls *.xlsx);;'
+                                                     '%s *.ods *.xls *xlsx;;'
                                                      '%s *.csv;;%s *.sav *.zsav *.por;;'
-                                                     '%s *.sas7bdat *.xpt;; %s *.dta;; %s *.rdata *.rds'%
-                                                     (_('All importable data files'), _('Text files'),
+                                                     '%s *.sas7bdat *.xpt;; %s *.dta;;'
+                                                     '%s *.rdata *.rds'
+                                                     %
+                                                     (_('All importable data files'), _('Spreadsheet files'), _('Text files'),
                                                       _('SPSS data files'), _('SAS data files'), _('STATA data files'),
                                                       _('R data files'))
                                                      )[0])
-        #*.txt *.log *.tsv
 
 
 def open_demo_data_file():
     return str(QtWidgets.QFileDialog.getOpenFileName(None, _('Open data file'), os.path.dirname(csc.__file__) +
                                                      '/sample_data',
-                                                     '%s (*.csv *.sav *.zsav *.por *.sas7bdat *.xpt *.dta *.rdata *.rds);;'
+                                                     '%s (*.csv *.sav *.zsav *.por *.sas7bdat *.xpt *.dta *.rdata '
+                                                     '*.rds *.ods *.xls *.xlsx);;'
+                                                     '%s *.ods *.xls *xlsx;;'
                                                      '%s *.csv;;%s *.sav *.zsav *.por;;'
-                                                     '%s *.sas7bdat *.xpt;; %s *.dta;; %s *.rdata *.rds'%
-                                                     (_('All importable data files'), _('Text files'),
+                                                     '%s *.sas7bdat *.xpt;; %s *.dta;;'
+                                                     '%s *.rdata *.rds'
+                                                     %
+                                                     (_('All importable data files'), _('Spreadsheet files'), _('Text files'),
                                                       _('SPSS data files'), _('SAS data files'), _('STATA data files'),
                                                       _('R data files'))
                                                      )[0])